🎨 chore: change environment for ver.2
