+++ old/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package dev.spring.petclinic.old.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
