+++ old/findowner/model/Pet.java
@@ -1,0 +1,14 @@
+package dev.spring.petclinic.old.findowner.model;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Builder
+@Getter
+public class Pet {
+    private int id;
+    private int ownerId;
+    private String name;
+}
