+++ src/main/java/dev/spring/petclinic/domain/Owner.java
@@ -1,0 +1,17 @@
+package dev.spring.petclinic.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Owner {
+    private Integer id;
+    private String firstName;
+    private String lastName;
+    private String address;
+    private String city;
+    private String telephone;
+}
