+++ src/main/java/dev/spring/petclinic/domain/PetType.java
@@ -1,0 +1,13 @@
+package dev.spring.petclinic.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PetType {
+    private Integer id;
+    private String name;
+}
