--- src/main/java/dev/spring/petclinic/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package dev.spring.petclinic.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
