--- src/main/java/dev/spring/petclinic/findowner/model/Pet.java
@@ -1,14 +1,0 @@
-package dev.spring.petclinic.findowner.model;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDate;
-
-@Builder
-@Getter
-public class Pet {
-    private int id;
-    private int ownerId;
-    private String name;
-}
