--- src/main/java/dev/spring/petclinic/searchResult/dto/OwnerDto.java
@@ -1,18 +1,0 @@
-package dev.spring.petclinic.searchResult.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Builder
-@Getter
-public class OwnerDto {
-    private int id;
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String city;
-    private String telephone;
-    private List<PetDto> pets;
-}
