+++ src/main/resources/mapper/OwnersMapper.xml
@@ -3,8 +3,8 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="dev.spring.petclinic.searchResult.mapper.OwnersMapper">
-    <select id="selectOwnerById" parameterType="int" resultType="dev.spring.petclinic.searchResult.model.Owners">
+<mapper namespace="dev.spring.petclinic.old.searchResult.mapper.OwnersMapper">
+    <select id="selectOwnerById" parameterType="int" resultType="dev.spring.petclinic.old.searchResult.model.Owners">
         SELECT * FROM owners WHERE id = #{id}
     </select>
 </mapper>
