+++ src/main/resources/mapper/VisitsMapper.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="dev.spring.petclinic.searchResult.mapper.VisitsMapper">
-    <select id="selectVisitsByPetId" resultType="dev.spring.petclinic.searchResult.model.Visits">
+<mapper namespace="dev.spring.petclinic.old.searchResult.mapper.VisitsMapper">
+    <select id="selectVisitsByPetId" resultType="dev.spring.petclinic.old.searchResult.model.Visits">
         SELECT * FROM visits WHERE pet_id = #{petId}
     </select>
 </mapper>
